fix: ライトモードの視認性を改善しコンポーネントを責務ごとに分割
- 背景色を #f8fafc に変更し白の眩しさを軽減
- コードブロックにボーダーを追加し背景色を調整 (#eff1f3)
- サイドバー・ヘッダーの背景色をコードブロックと統一
- ヘッダーを全幅レイアウトに変更しロゴ・favicon を配置
- app.tsx / preview.tsx / file-tree.tsx / quick-open.tsx 等を
  コンポジションパターンで責務ごとにコンポーネント分割
- Prism テーマ選択を usePrismTheme フックに共通化

Closes #53
