+++ src/client/components/file-tree.tsx
@@ -1,16 +1,14 @@
 import type { FileNode } from "@shared/types";
 import {
-  Search,
-  ChevronRight,
   ChevronDown,
+  ChevronRight,
+  FileText,
   Folder,
   FolderOpen,
-  FileText,
 } from "lucide-react";
-import { useState, useMemo, useCallback, useEffect, useRef } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 
-import { Input } from "@/components/ui/input";
-import { Kbd } from "@/components/ui/kbd";
+import { FileTreeSearch } from "@/components/file-tree-search";
 import { Tree, TreeItem, TreeLeaf } from "@/components/ui/tree";
 import { computeAutoExpandPaths, findNode } from "@/utils/auto-expand";
 
@@ -187,23 +185,7 @@ export const FileTree = ({ files, selectedPath, onSelect }: FileTreeProps) => {
 
   return (
     <div>
-      <div className="relative mb-3">
-        <Search
-          size={14}
-          className="absolute left-2 top-1/2 -translate-y-1/2 text-muted-foreground"
-        />
-        <Input
-          type="text"
-          value={query}
-          onChange={handleQueryChange}
-          placeholder="Search files..."
-          className="pl-7 pr-12 py-1 text-sm"
-        />
-        <div className="absolute right-2 top-1/2 -translate-y-1/2 flex items-center gap-0.5 text-xs text-muted-foreground">
-          <Kbd>⌘</Kbd>
-          <Kbd>P</Kbd>
-        </div>
-      </div>
+      <FileTreeSearch query={query} onChange={handleQueryChange} />
       <Tree>
         {filtered.map((node) => (
           <FileTreeNode
