+++ src/client/components/markdown-table.tsx
@@ -1,0 +1,10 @@
+import type { ComponentPropsWithoutRef } from "react";
+
+export const MarkdownTable = ({
+  children,
+  ...props
+}: ComponentPropsWithoutRef<"table">) => (
+  <div className="overflow-x-auto">
+    <table {...props}>{children}</table>
+  </div>
+);
