+++ src/client/hooks/use-prism-theme.ts
@@ -1,0 +1,8 @@
+import { themes } from "prism-react-renderer";
+
+import { useTheme } from "@/hooks/use-theme";
+
+export const usePrismTheme = () => {
+  const { theme } = useTheme();
+  return theme === "dark" ? themes.vsDark : themes.github;
+};
